requirements.txt
python-dotenv
discord
matplotlib
GitPython